Fix memory write test
Poking a wire with a value seems to set it permanently to that value until it is
poked again with a new value.

In this case the memory write test had to
be fixed by setting memwrite to false after all of the writes, so that
the test doesn't iterate through the remaining addresses in RAM and
overwrite them with writedata's retained value of 355.