Implement asynchronous memory module
This new memory module has a configurable latency that delays the
memory's operations for a certain amount of clock cycles. This
effectively models real DRAM, which often takes hundreds of cycles for a
load or store operation to go through and necessitates adding SRAM caches
to a system.

However the memory module isn't working properly yet - either the write
tests for it are written wrong or memory writing is bugged