Convert PartialWrite to OutstandingReq
This treats reads and write the same at the port interface. Now, you
also store the address, etc. when there is an outstanding read and use
this data when the response returns.

Note: There's a nasty edge case if you have write followed by something
else. You cannot send the next request (read) until after one cycle so
the write part of the write request can be sent. I don't think the tests
test for this.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>