Move branch predictor CPU into its own file
Revert pipelined/cpu.scala and components/hazard.scala to 4699e67 and
fix up the other minor changes. This also creates two new files
pipeliend/cpu-bp.scala which contains the CPU design with a branch
predictor and components/hazardbp.scala which contains the updated
hazard detection logic with an extra IO.

Also update the logic to instantiate the CPU to understand the
difference between the base pipelined CPU and the branch predictor
version of the CPU design. This required updating the tester driver and
the CPU tests as well.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>