Add ready signal to stable interface
This signal serves as a way for the pipeline to know whether the backing memory
is ready for another request. This is required for the non-combinational memory
to stall when the memory is not ready to continue.

Previously, the 'good' signal served this purpose but is now used to write valid
memory data to each stage's register.