The width of PC of single cycle cpu is left undefined (#150)
This means FIRRTL compiler will infer the bit width of PC. It is shown in the
verilog presentation that FIRRTL optimized the PC width to 5. That results
in the single cycle CPU to not to cause any error for small application.
However, when run with large applications, the PC reached 0x1ffc and looped
back to 0 for the next PC.

This patch, by explicitly defining the width of PC, prevents the FIRRTL complier
from over-optimizing the width.

Essentially, width inferences cause design issues. Don't do it.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>