Add newly generated binary supporting the dual-issue cpu
Basically, we pad more nop at the end of the binary so that the
program won't terminate early.

Note: the simulation ends when the value of the PC register is
higher than the address of the `_last` symbol in the binary
(use readelf tool to find where the `_last` symbol). This means,
in a 5-stage pipeline, if the PC of the issue stage is larger
than the `_last` symbol, the simulation stops. Consequently,
for the last meaningful instruction to reach the writeback stage
before the simulation ends, we need to add at least 10 nop's
before the `_last` symbol.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>