Merge dual issue cpu from winter quarter 2022 (#153)
* Add newly generated binary supporting the dual-issue cpu

Basically, we pad more nop at the end of the binary so that the
program won't terminate early.

Note: the simulation ends when the value of the PC register is
higher than the address of the `_last` symbol in the binary
(use readelf tool to find where the `_last` symbol). This means,
in a 5-stage pipeline, if the PC of the issue stage is larger
than the `_last` symbol, the simulation stops. Consequently,
for the last meaningful instruction to reach the writeback stage
before the simulation ends, we need to add at least 10 nop's
before the `_last` symbol.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>

* Make instruction responses 8-byte aligned

Instruction ports now response with 64 bits of instructions.
Specifically, they are instructions at PC/8 and PC/8+4.

This is for supporting dual-issue, where the cpu must be able to receive
64 bits of instructions in each cycle. This leads to,
- Widened instruction response/receive port.
- Additional logics for single cycle cpu and pipelined cpu to issue the
correct instruction.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>

* Add dual-issue cpu

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>

* Revert the design of non-combinational memory CPU

Signals changed, and as a result, the tests are broken due to
the non-combinational CPU. Will fix this later.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>

* Merged components that are used by both pipelined and pipelined-dual-issue cpus

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>